* Nombre del Proyecto: Mi Monte Fase 2 - Conciliación * Se actualiza script y se corrige metodo guardado de conciliacion.

+++ src/main/java/mx/com/nmp/pagos/mimonte/dao/conciliacion/SubEstatusConciliacionRepository.java
@@ -22,7 +22,7 @@ import mx.com.nmp.pagos.mimonte.model.conciliacion.SubEstatusConciliacion;
  * @version 0.1
  */
 @Repository("subEstatusConciliacionRepository")
-public interface SubEstatusConciliacionRepository extends JpaRepository<SubEstatusConciliacion, Integer> {
+public interface SubEstatusConciliacionRepository extends JpaRepository<SubEstatusConciliacion, Long> {
 
 	/**
 	 * Búsqueda del sub estatus de la conciliación a partir de la descripcion.
